requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
requests
django-extensions